fermo al video 86 - Sfida - Visualizzazione delle tabelle collegate con la ricerca per il barcode , aggiunta la selezione delle altre 2 tabelle nel repository nel metodo bycodice funzionante , aggiunto anche negli altri 2 metodo selbydescrizione e selbybarcode attualemnte non va da sistemare , errore quando si recuperano i dati , nel controller aggiunte righe necessarie per la creazione dei nuovi attributi , aggiunti 1 nuovo dto , 1) ivaDto , categoria e solo 1 campo di tipo stringa prelevato dalla tabella famAssortimento

+++ salutiWebApi/salutiWebApi/Models/Iva.cs
@@ -10,7 +10,7 @@ namespace salutiWebApi.Models
     public string? Descrizione { get; set; }
 
     [Required]
-    public int Aliquota { get; set; }
+    public Int16 Aliquota { get; set; }
 
     //relazione : 1 iva corrisponde a molti articoli
     public virtual ICollection<Articoli>? articoli { get; set; }
